MAJOR DIRTY: significant breaking changes to how parameters are handled
no parameters are hardcoded in the Parameters object anymore (except for
a few paths for now)

all simulation-relevant parameters are defined in the lua config file
and read into the program. if the param is constant, its value is
immediately loaded; otherwise, it expects the value to exist in the
database and to be read in before a simulation starts.

individual parameters are stored as a Parameter and the Parameters
object stores all of these individual params and allows access through
the get() method.

the parameters object also calls the database read itself to ensure that
all of the values it expects to have values for it eventually does.

params can be retrieved either from the fullname or nickname defined in
the lua config file.

TODO: code cleanup, documentation (and prob some optimization to improve
initialization speeds --- but i need to time how long each step in a
full job takes)

TODO: calc_strain_probs() is a very fragile part of the object as it
needs to be called any and everytime param values are read (from the
config or the database).