CLEAN: reimplement copy parameters and cleaned up parameter loading
because all parameters are saved to the database (const, step, and
copy), during reading, we can just look for all of the names we expect
to find. this doubles as a safety check to ensure we read the correct
number of parameters as expected. to do this, we must manually handle
the case of the seed parameter because this is not specified in the
config file but is automatically added by the program.